36 коммит. Программа работает корректно

+++ src/main/java/ru/skypro/homework/dto/comments/CommentDTO.java
@@ -18,7 +18,7 @@ public class CommentDTO {
     private String authorFirstName;
 
     @Schema(description = "дата и время создания комментария в миллисекундах с 00:00:00 01.01.1970")
-    private int createdAt;
+    private long createdAt;
 
     @Schema(description = "id комментария")
     private int pk;
